Rename and reposition specialist comment presenter
