--- engines/bops_api/app/presenters/bops_api/v2/specialist_presenter.rb
@@ -1,29 +1,0 @@
-# frozen_string_literal: true
-
-module BopsApi
-  module V2
-    class SpecialistPresenter
-      extend Forwardable
-      def_delegators :@consultee, :id, :organisation, :role, :email_sent_at, :planning_application_constraints
-
-      def initialize(consultee, responses)
-        @consultee = consultee
-        @responses = responses
-      end
-
-      def reason
-        active_constraints = planning_application_constraints.active
-        reason_type_code = active_constraints.first&.constraint&.type_code
-        reason_type_code.present? ? "Constraint" : "Other"
-      end
-
-      def active_constraints
-        planning_application_constraints.active
-      end
-
-      def comments
-        @responses
-      end
-    end
-  end
-end
